Remove nondeterminism from runs of the refcnt test. [reviewed by vass]
The nondeterminism in the output of users/ferguson/refcnt.chpl:
  writeln("Before scope (begin)");
  sync {
    var r:R;
    begin {
      test_one(r, false);
    }
  }
  writeln("After scope (begin)");
was due to calls to chpl__autoDestroy() being inserted after the _downEndCount
call at the end of a begin block.  The autoDestroy calls belong to the "begin"
block, so their actions should be completed before the "begin" block declares
that it is done.  If this condition does not hold, then those actions can "leak
out" of a surrounding sync block.  (And indeed, that is what was being
observed.)

In particular, the destructor for the record type R used in the test calls
writeln, so there is observable behavior associated with when the autoDestroy
call actually runs.  Code following the sync block is enabled as soon as
_downEndCount() is executed.  This introduces a race between that code and and
code which follows the _downEndCount call in [code that was generated from] the
begin block.  So "After scope (begin)" might be printed before the message
indicating that the r inside the sync block is being destroyed.

For the fix, I just forced autoDestroy calls added to a begin block to be
inserted before the call to _downEndCount, rather than after.  Along the way, I
found I needed to flatten the body of a begin block when it is converted to a
function.  

The assumption that every block ends with an executable statement turned out to
be false for modules.  But the conversion in callDestructors is only intended
to deal with blocks inside of functions, so I just filtered out that
troublesome case (module blocks) at the top of the loop.  


  DETAILS

    compiler/include/symbol.h
    compiler/AST/symbol.cpp
Added a new utility FnSymbol::insertBeforeDownEndCount(Expr*).

    compiler/passes/createTaskFunctions.cpp
Flatten the body of a begin block when it is turned into a function.

    compiler/passes/parallel.cpp
Use insertBeforeDownEndCount() instead of insertBeforeReturnAfterLabel() to
place the inserted autoDestroy calls ahead of that _downEndCount() call.

    compiler/resolution/callDestructors.cpp
In insertAutoDestroyCalls(), filter out blocks that are the children of a
module.  Add calls to "_downEndCount" to the list of special cases that can
terminate a block ahead of its last statement.

    test/users/ferguson/refcnt.execopts
Removed.  This was an earlier but ineffective attempt to resolve the
nondeterminism.


[tested on refcnt.chpl with both standard and --no-local]
[tested on domains/ distributions/ multilocale/ performance/ with comm=gasnet]
[testing on test/ using <std> (in progress)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22461 3a8e244f-b0f2-452b-bcba-4c88e055c3ca