Beefing up a test, enabled by r22438, in test/parallel/taskPar/taskIntents.
When the test was created, I had to work around a bug (see
  test/parallel/begin/vass/begin-dom-arr-mem-corrupt.chpl
and its .future as of r22166). I did that using .patch files in src/.

Now that the bug is solved, the workaround is no longer necessary,
so I am deleting the .patch files and updating .chpl files,
which result from running 'make clobber all' in src/.

The support for .patch files in src/Makefile is now not needed,
but I am not bothering to remove it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22463 3a8e244f-b0f2-452b-bcba-4c88e055c3ca