An update for TRIAGE/REGRESSIONS for this week.
* The refcnt test should now succeed, I removed its entry.

* Massive reduction in core dumps for linux64.pgi.none.fifo.pthreads.flat
on chpbld02, after six weeks of no testing logs:

11/04: total errors: 58;  "signal 11" crahses: 61
12/15: total errors: 15;  "signal 11" crahses: 12

I did not document this because I did not see how.

* weirdcase.chpl stopped timing out on gasnet-everything
I did not document this because I did not see how.

* A couple of nondeterministic segfaults under gasnet.qthreads
- recorded in TRIAGE

* hello3-datapar stopped timing out on gasnet-qthreads.
I updated REGRESSIONS.

* parallel/begin/vass/begin-dom-arr-mem-corrupt should no longer be
a passing future, starting 12/20.

* I removed the entry for simple-alias-mem-use failure from TRIAGE.
I do not see a reason to suspect that it failed recently - there are
no log files for these gnu/pgi/intel configurations on chpbld02
between 11/5 and 12/14.

* Removed from REGRESSIONS "CCE INTERNAL error" due to a CCE bug -
they were gone on 12/20 on chpbld01 for:
   cray-xc.cray-prgenv-cray.none.fifo.pthreads.flat--linux64.gnu
Still leaving an entry in TRIAGE - watch for
  Nightly/weekend/day6-Sat-cray-xe.cray-prgenv-cray.none.fifo.pthreads.flat--linux64.gnu-std.log

* Removed from TRIAGE the entry for refcnt - Tom fixed as of 12/20.

* Five random baseline compilation timeouts - in TRIAGE.

* spec/Classes/defineMethod regression has been fixed - removed from
REGRESSIONS.

* Recorded in REGRESSIONS three gasnet tests that fail with increased
comm counts, believed due to r22438. There is an email discussion
of what to do about those.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22470 3a8e244f-b0f2-452b-bcba-4c88e055c3ca