Cleaned up noRefCount-related code in modules/internal.
[reviewed by Tom]

Clarified the behavior that when noRefCount=true,
no reference counting or dist/domain/array deletion occurs.

This is based on code examination and common sense,
and discussed within the group.

* Added compilerAssert(!noRefCount) to incRefCount(), decRefCount(),
destroyDist(), destroyDom(), destroyArr(),
chpl_incRefCountsForDomainsInArrayEltTypes(),
chpl_decRefCountsForDomainsInArrayEltTypes().

* As a side effect, atomic operations on the reference counts are now
not performed at all when noRefCount=true. This should be beneficial
for performance in this mode when --no-local.

* Ensured that when noRefCount=true, the destructors for _distribution,
_domain, _array do nothing. Removed inner checks for !noRefCount.

While there...

* Moved the calls to remove_dom() and remove_array() from
destroyDist() and destroyDom(), for clarity.

* Wrapped those calls in 'local', to avoid 'on' clauses
when dealing with atomics in the callee.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22473 3a8e244f-b0f2-452b-bcba-4c88e055c3ca