Slight simplification to yesterday's commit
[trivial, no review]

In adding some comments before committing as requested by Tom
yesterday, I realized that for unary operations, the result
type is easily inferred in both the homogeneous and heterogeneous
case, so collapsed these into a single, loop-based form rather
than the previous single, lisp-y style forme.   Added a test
to verify that types are preserved properly where the behavior
is unchanged since prior to this change.

(in writing up this test, am noting that we don't maintain
boolean widths in doing operations on booleans, as with other
types, which now seems inconsistent.  Will work on a future
for this behavior).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22476 3a8e244f-b0f2-452b-bcba-4c88e055c3ca