Don't run single-thread fifo tests on > 1 locale, because they always fail.
[Not reviewed.]

The tests QueuedTasks, OneThread, and OneThreadSyncStmt were all written
specifically to test single-thread behavior with tasks=fifo (which used
to be called tasks=pthreads).  They can't work properly in multilocale
execution with tasks=fifo nowadays, because the r22232 optimized locales
initialization using coforall+on needs multiple tasks per locale and we
can't run multiple tasks per thread with fifo.  Therefore, turn these
tests off for CHPL_COMM!=none.  They still fulfill their original
intent, which was to test the fifo tasking layer, but only on a single
locale.

To offset this restrictiopn, for OneThreadSyncStmt, rather than skipping
when tasks!=fifo, skip when tasks=none.  That test does work for all the
tasking implementations other than none, so there's no reason not to
test it for all of them.  This is especially true when for example the
major benefit of qthreads is user task switching on threads, or at least
shepherds; we might as well test that.

While here, I also added a skipif for tasks==none for OneThread.  It
didn't use to have one even though it was originally written to test
tasks=fifo just like OneThreadSyncStmt was.  It doesn't work with
tasks==none any better than the others do.

I didn't change the CHPL_TASKS!=fifo skipif for QueuedTasks.  It will
potentially produce different output for each tasking implementation,
and I didn't want to go to the trouble of figuring out all the cases at
this time.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22492 3a8e244f-b0f2-452b-bcba-4c88e055c3ca