Update comm counts due to r22438.
Revision 22438 added autoCopy and autoDestroy calls for user-defined record
types (passed by value) and internally reference-counted types, for data items
shared across a nonblocking "on".  

The autoCopy calls are necessary, to ensure that shared data is not released
prematurely by the spawning task.  The autoDestroy calls are then necessary to
release the shared data when the spawned routine is done.  (This manipulation
of reference counts was already in place for "begin" blocks; r22438 merely
expanded it to include nonblocking "on" blocks as well.)

This patch contains updates to the communication counts resulting from the
change.  Verbose logging was used to trace the changed comm counts directly
back to the changes introduced in r22438.  The changes are expected and easily
explained in terms of the added autoCopy and autoDestroy calls, as detailed
below.

DETAILS

    test/modules/sungeun/init/printInitCommCounts.good
One more fork is executed on each remote locale.  The extra remote call is due
to the inserted autoDestroy, which releases the "flags" array used to
synchronize initialization of the Locales array in initOnLocales().  The added
autoCopy calls, used to retain the "flags" array for the sake of each remote
locale do not add any communication, because the calls and the "flags" array
are all local to the root locale.

    test/studies/ssca2/performance/SSCA2_commDiags.good
The fork count on each remote locale increased by 1, and the get count
increased by 2.

The extra forks are due to inserted autodestroys for two shared variables:
_coforallCount and _formal_tmp_heavy_edge_list.  Only one of these generates a
fork, because the first variable is an integer and therefore its autoDestroy is
trivial.  

The extra gets are because the variables are read before being destroyed.
These reads can probably be removed as an optimization.  The wide access
inserts a read because the argument to autoDestroy is passed by value.  If its
intent was ref instead, the extra gets would go away.

    test/studies/hpcc/CommDiags/stream-ep_commDiags.good
The fork count increased by 2 on each remote locale.  Both are due to
autoDestroy calls added by r22438.

    test/performance/hilde/coforallOnByRef.chpl
    test/performance/hilde/coforallOnByRef.numlocales
    test/performance/hilde/coforallOnByRef.bad
    test/performance/hilde/coforallOnByRef.good
    test/performance/hilde/coforallOnByRef.future
Added a future to capture Vass's suggestion that extra comm counts can be
eliminated for the special case of a noblocking "on" inside a coforall.

    test/performance/hilde/userSync.chpl
    test/performance/hilde/userSync.numlocales
    test/performance/hilde/userSync.bad
    test/performance/hilde/userSync.good
    test/performance/hilde/userSync.future
Added a future to capture the fact that the "no copy" and "no auto destroy"
flags are not propagated through temp variables.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22495 3a8e244f-b0f2-452b-bcba-4c88e055c3ca