Add new compiler options --explain-call-id and --explain-verbose.
--explain-verbose is a user option that turns on the disambiguation trace for
the selected function (using --explain-call) or instantiation (using
--explain-instantiation).  

--explain-call-id is a developer option that selects a call to explain using
the call ID.

To get the trace output to be properly interleaved, I changed the
TRACE_DISAMBIGUATE_BY_MATCH(...) macro to write its output to stderr (like
the output from --explain-call, etc.)

TODO: In this patch disambiguation tracing is always enabled at compile time.
We need to review compilation times, and see if this is acceptable.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22497 3a8e244f-b0f2-452b-bcba-4c88e055c3ca