Add two new versions of fasta: fasta-lines and fasta-printf
fasta-lines builds up a string for each full line of output within the
randomMake procedure, rather than writing each character out one by one. The
string concatenation causes a very large amount of memory to be allocated and
leaked, but it is still faster than fasta.

fasta-printf performs some trickery to use arrays as character buffers and
hands it straight to printf, bypassing qio. It is just meant to show the
overhead we currently have with string manipulation and qio routines.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22498 3a8e244f-b0f2-452b-bcba-4c88e055c3ca