Improve discussion of env vars that control Cray specific runtime behavior.
[Not reviewed.]

In the part of README.cray that talks about parameterizing the
Cray-specific runtime layers, add introductory information about the
default settings, and swap the subsections that talk about the ugni comm
layer and the muxed tasking layer.  Since the comm layer discussion
refers to one of the parameters that affect the tasking layer, tasking
before comm is the better order.

(This fell out of discussions back in fall 2013, with a Cray internal
user who was doing some Chapel performance studies and ended up chasing
the appropriate environment settings unnecessarily.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22508 3a8e244f-b0f2-452b-bcba-4c88e055c3ca