Minor tweaks to start_test/sub_test
[reviewed by Tom (start_test) and Kyle (sub_test)]

start_test: if "make spectests" and/or "computePerfStats" fail,
make their logs available - and announced in start_test's output.

sub_test: if the environment environment NUMLOCALES is not set,
treat it as if it is set to 0.
  In this case, a '-nl' option will not be passed to the executable
in the absence of an applicable numlocales/NUMLOCALES file.
start_test always sets NUMLOCALES, so this is just a safety measure.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22509 3a8e244f-b0f2-452b-bcba-4c88e055c3ca