Don't run single-thread fifo tests with --no-local, because they always fail.
[Not reviewed.]

This is a follow-up to r22492.  Don't run QueuedTasks, OneThread, and
OneThreadSyncStmt when they are compiled with --no-local, because they
hang in that configuration just as they do with CHPL_COMM!=none.

(These are still perfectly valid tests, because they actually test our
tasking behavior, not anything to do with our multilocale behavior.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22523 3a8e244f-b0f2-452b-bcba-4c88e055c3ca