Remove log button from graphs with data points of 0 or less [reviewed by Tom]
I noticed that when data points were present of 0 or less, the graphs generated
would lose data points when the log button was pressed, and if one had zoomed in
vertically before pressing the log button, the data lines were lost entirely.
This occurred because dygraphs does not support log on data points of 0 or less.
To avoid later confusion, I altered our graph generation such that a log button
would be colored red in the case where a graph would contain unsupported data
points, since it seems reasonable for us to want to utilize the log
functionality over the portion of the graph that is loggable but there should be
some form of warning mechanism otherwise.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22531 3a8e244f-b0f2-452b-bcba-4c88e055c3ca