Minor wording updates to CHPL_*_COMPILER section
[reviewed by Kyle]

Related to Kyle's commit r22537, we were wrestling with how to improve
this section of the README to more clearly/accurately describe the
situation in which CHPL_TARGET_COMPILER will inherit from
CHPL_HOST_COMPILER.  This is what we came up with.  While here,
re-alphabetized the platform entries, distinguished cray-x* from
cray-*, and tightened up the text introducing this section slightly.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22540 3a8e244f-b0f2-452b-bcba-4c88e055c3ca