Set OPTIMIZE and DEBUG in the third-party directory as we do in the compiler and runtime
[reviewed by Lydia]

qThis commit puts logic in the top-level third-party Makefile that sets
OPTIMIZE and DEBUG based on the setting of CHPL_DEVELOPER as we
already do for the compiler and runtime directories.

I've also removed that same logic from the re2 Makefile as it is no
longer needed, assuming you call make from at least one level up.

So that is a caveat: you do not get this behavior if you go into a
specific third-party directory and do 'make'.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22545 3a8e244f-b0f2-452b-bcba-4c88e055c3ca