Style-oriented housecleaning for tests looping over domains unnecessarily
[reviewed by lydia, kyleb, sungeun, hilde, diten, rafa]

This patch does a cleanup over the test directory, looking for cases
where a loop is looping over a domain expression unnecessarily.  An
example of this is:

        for i in {1..n} do

This says "Create the anonymous domain {1..n} and then loop over
it. Since domains are heavier-weight than ranges, the preferable form
in an optimized compiler is:

        for i in 1..n do

which only creates an anonymous range.  

Of course, ultimately, we'd like the compiler to strength-reduce such
cases, but until then we have to do it manually.  Even if we did have
the optimization, I think we want to put forth codes that demonstrate
preferred Chapel style, so rewriting these cases is beneficial anyway.

Note that there are still 3-5 cases that use the old forms because
part of their point in being tests seems to be to compare the
different forms.

One of the cases was in an internal module as well.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22549 3a8e244f-b0f2-452b-bcba-4c88e055c3ca