New fannkuch versions
We reviewed fannkuch in the performance meeting today, and this
is me having some fun writing it more in my Chapel style and
trying some variations.  I added these to the graph file to
compare against Kyle's going forward.

The versions are:

fannkuch-redux-blc.chpl : my main version; relative to Kyle's
   version, uses int(64)s, type inference, declare-when-first-
   defined style (moreso), and more Chapel-isms (loop mechanisms,
   swap operator, etc.

fannkuch-redux-int8.chpl : this is a version in which I tried
   using int(8)s to see if the more compact memory footprint
   would help performance much.  Answer: not from what I've
   seen.  This version is also parameterized so that one can
   compare different int formats with a single source.

fannkuch-redux-compact.chpl : this version is terser, but
   probably also slower by using data parallelism in cases
   that don't warrant it.  This raises an increasingly common
   question: How can we have people write things cleanly without
   getting parallelism they don't want/need.


And one future:

fannkuch-redux-blc-unbounded.chpl : this is a future that tries
   to do 'for i in 1.. { ... }' to get an unbounded loop (that
   it will break out of before we run off the end of the unbounded
   range).  I wasn't entirely surprised that this didn't work, but
   I was surprised that it resulted in a syntax error rather than
   a runtime or compile time complaint about iterating over an
   unbounded range.  So this is a feature request.

(When this future is retired, it can be folded into the other
versions and removed rather than maintaining a fourth version.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22552 3a8e244f-b0f2-452b-bcba-4c88e055c3ca