Allow casts between c_ptr types and allow c_ptrTo on arrays.
[reviewed by Tom]

This patch enables casts (by providing a _cast proc) for
c_ptr types and also creates a special version of c_ptrTo
for arrays that checks for a 1-D rectangular arrayand
returns a pointer to the first element.

I also changed the name of the extern function c_ptrTo
to _c_ptrTo so that c_ptrTo is the version with the extra
checks (that normally just calls c_ptrTo which just returns
the original pointer anyway).

Future work: better error handling for c_ptrTo and doc updates.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22574 3a8e244f-b0f2-452b-bcba-4c88e055c3ca