Temporary measure to help shift SVN root for our test machines
This commit temporarily removes the read-only copy of the SVN
tree that's updated each night for nightly testing before
updating it in order to support attempts to move to a new SVN
tree without having to log into and clean each test machine.
Once we've run on each machine, we should be able to remove
this command and return to the "simply update" model.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22581 3a8e244f-b0f2-452b-bcba-4c88e055c3ca