Modified .good to accept the no-local regression in printFinalMemStat.chpl due to r22473.
[As discussed with the group.]

To summarize the cause of the regression, r22473 added a 'local' clause
in destroyDom() and destroyArr(). Against my expectation, that did not
cause the compiler to eliminate the 'on' clauses in the functions
remove_dom() and remove_arr() that are invoked from there. However,
independently of r22473, the 'on' argument bundle gets an extra pair of
(line number, file name) arguments when its enclosing function is
compiled in both 'local' and non-'local' modes. These extra arguments
in 4 bundles are what causes the extra 64 bytes memory usage witnessed
by this test. 2 of those bundles are allocated concurrently, bumping
the maximum memory usage by 32 bytes.

See the thread "r22473 and compiler issues" on the internal mailing list
for a more detailed description of the above.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22586 3a8e244f-b0f2-452b-bcba-4c88e055c3ca