No longer remove successful performance test .tmp files
[concept reviewed by vass]

Lydia and I have recently run into confusion where we've
needed to validate if a performance test got the right
output, but the testing system has already cleaned up
after itself, so it's not possible to do so.  Ultimately,
we'd like a way to specify "match" performance keys that
must get a certain result for a performance test to pass,
but until that day, let's keep around the .exec.out.tmp
files whether the test "passes" or not so that we can
verify by hand.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22591 3a8e244f-b0f2-452b-bcba-4c88e055c3ca