Fix RE2 builds and tests on Mac OS X Mavericks
Even after upgrading RE2 to the latest version which addresses
some issues with Mac OS X, I was having trouble linking the C++
test programs with RE2 - and I believe that .chpl programs
using RE2 would not compile either. The source of the problem
was:
      1) Mac OS X does not support static linking of executables
         (although it can link with a static library), so using the
         --static flag in the tests is not an option.
      2) The Mac OS X linker does not support -rpath=something;
         you have to give it -rpath something.
      3) The RE2 build/install process leaves a partial pathname
         to the .so encoded in the .so; so that when you link with
         that binary it would look for the re2.so in an obj/ directory.
         Because of this partial pathname, the usual rpath linker
         flag was not enabling the loader to find the required library.
      4) When dynamically linking, the list of exported symbols
         did not include some that earlier patches to RE2 (to make
         it work on channels) added.

To fix these problems, this patch:
      1,2) Adjusts the C++ language Regexp tests to not use
           --static on Mac OS X and to use -rpath something

      3) Fixes the RE2 Makefile to update the pathname
      in the .so to the absolute destination pathname using
      the tool install_name_tool if it exists

      4) Adds the missing symbols to libre2.symbols and
      libre2.symbols.darwin.

This version passes test/regexp on Linux and has minimal changes from the one
that worked for me with Mac OS X Mavericks.

Note that for Mac OS X versions before Mavericks, the RE2
version currently in trunk does not compile and link properly.
Removing the symbol for operator<<(std::ostream&, re2::StringPiece const&)
from libre2.symbols.darwin solves the problem. We havn't done
that in our own patch to RE2 because we expect that the RE2
project will fix the problem in the relatively near future - but
we could make that change e.g. before the next release if necessary.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22593 3a8e244f-b0f2-452b-bcba-4c88e055c3ca