Stop treating {D} as D for a domain D
[reviewed by hilde]

For some time now, Chapel has permitted a domain to be specified either by
naming it, e.g.:

        const D = {1..n};
        const DistD = D dmapped Block(...);
        forall i in DistD ...

or by wrapping it in an extraneous set of curly brackets:

        const D = {1..n};
        const DistD = {D} dmapped Block(...);
        forall i in {DistD} ...

The reasons for this are somewhat obscure (and arguably indefensible), but
relate to (1) a precedent set in ZPL, (2) the days when we used square
brackets for domain literals, and (3) to the use of anonymous domains in
array declarations without curly brackets (i.e., a pure implementation of
Chapel would require:

        var A: [{1..n, 1..n}] real;

rather than permitting the shorthand:

        var A: [1..n, 1..n] real;


I'd forgotten about this anomaly until doing cleanup on iterations over
anonymous domains the other week when I encountered some instances of it.
Over lunch, Vass pointed out that cleaning this up should be fairly easy so
I took it on as a five-minute fix (TM).

This patch starts us toward a better world where we interpret {D} as an
associative domain of domain values, containing one index 'D'.  This seems
more consistent with the current language definition and was generally
agreed to be improved by the Chapel team.  I say "starts" because for now it
makes {D} into an error overload to help users with existing code transition
from the old world to the new.

The heart of the patch is a simple change:  We have traditionally supported
a routine, chpl__buildDomainExpr() which creates a domain from whatever is
passed it.  One of the overloads took a domain and simply returned it which
is what caused '{D}' to evaluate to 'D'.  So the idea behind this change is
to remove that overload (though again, for now, I simply replaced it with an
error overload while we're transitioning).

A little more work was required, however, because the same call is used to
build the domain for an array type like '[D] real' and in this case, we want
to permit a domain like 'D' to be used.  So what I did was to split this
call into two calls:


        chpl__buildDomainExpr() : does what it has, but making the domain
                                  case into an error overload for now

        chpl_ensureDomainExpr() : used in the array type signature context
                                  to make sure that the expression is a
                                  domain.  In this context, if the thing
                                  is a domain, we simply return it.  In
                                  all other cases, we forward the args to
                                  chpl__buildDomainExpr()

From there, the only other change was to switch calls to
chpl__buildDomainExpr() in the array type signature case (which was most of
them) over to chpl__ensureDomainExpr(), and to rewrite any tests that were
relying on '{D}' evaluating to 'D'.

For backwards compatibility, this commit generates an error message when
encountering {D} to tell the user about the change in interpretation that's
coming and to suggest that they remove the curly brackets.  In the next
release, by removing the first overload of chpl__buildDomainExpr() in
ChapelArray.chpl, we'll have the new interpretation.  Note, however, that we
don't yet support associative domains of domains, so disabling this overload
today results in a "we don't support this style of associative domain
yet"-style message.  But I think after the next release, we should still
remove this overload even if we haven't yet gotten around to fleshing out
associative domains of domains (which might be as simple as providing the
required hash function for domain values?)

I also added tests to lock in the error message for now.  Ultimately, these
could be used to test the associative domain of domains intepretation.

The bulk of the patch is just updating tests that had relied on this
syntactic form in the past.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22596 3a8e244f-b0f2-452b-bcba-4c88e055c3ca