Add support for expression-less 'serial' statement
[reviewed by lydia]

This commit extends the parser to permit 'serial do ...' or 'serial { ... }'
without the requirement to add an expression test.  The meaning is the
same as 'serial true ...'  In retrospect, I don't know why we didn't do
this years ago (maybe we didn't care enough).

Updated the spec to reflect this change and wordsmithed the serial
statement description while here.  Updated a 'serial true' test in
the spec to simply use 'serial'.

Updated my code that motivated this 5-minute fix (TM).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22598 3a8e244f-b0f2-452b-bcba-4c88e055c3ca