Add a perf test for emitted code size [reviewed by ]
This adds a simple perf test and graph for emitted code size.

I added a counter in codegenStmt(), so the number of statements generated is
tallied.  This is printed out conditionally when fPrinEmittedCodeSize is true,
which is effected by tossing the --print-emitted-code-size compiler option.

The actual number of lines of C code generated will be somewhat larger, since
this count does not record the contents of the header file or function
signatures.

[Smoke-tested with hello.chpl]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22600 3a8e244f-b0f2-452b-bcba-4c88e055c3ca