Removing a stray over-curlied {D}
This .future contained an instance of {D} where D was needed.
Removing the {D} restores it to its prior behavior before my
change.  Scanning the nightly test logs, it appears that this
is the only .future that mistakenly had this condition (two
others used {D} with the intention of it meaning "associative
domain of domains".



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22603 3a8e244f-b0f2-452b-bcba-4c88e055c3ca