Bug fix for LICM related to a variable aliasing a record's field
[reviewed by tom]

When the field of a record is deffed, we need to mark a def of the record
as well. This was handled for regular definitions of records, but not for defs
of variables that alias a record's field. 

For instance in the IR: 

call_tmp = &(record.field) 
func(call_tmp) 

was recognized as a def to call_tmp and field, but not to record. 

This patch updates the def analysis so that defs to variables that alias a
record's field are marked as defs to the record as well. 

This bug was exposed through Tom's work with the new assignment signature
(assignment using ref intents for the lhs instead of using PRIM_MOVE.) Prior to
this aliases to the field of a record were not generated in the AST in any test
cases (although they could have been in a user defined equals function.) 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22606 3a8e244f-b0f2-452b-bcba-4c88e055c3ca