Added an example with conditional statements to the spec.
[reviewed by David]

This is in response to a user's suggestion for the documetation
to reflect Chapel's ability to do if ... else if ...

I thought the best way to do so is by having it in an example.
However, it wasn't easy to weave such an example in, so I played
with the surrounding text: added a sentence that if ... else if ...
is allowed and moved the "each statement has its own scope" sentence
so it is not being a poor relative at the end of the section.

I also included a "hanging if" into the example.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22615 3a8e244f-b0f2-452b-bcba-4c88e055c3ca