Adding a performance future for coforalls over the locales array. [reviewed by Sung]
As discovered in BlockDist.chpl, iterating over the locales array in this manner
ends up creating a task per locale, as it should.  Our performance would improve
if in this case a task was not created for the originating locale, as it will be
able to just pick up the creating task for its purposes.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22620 3a8e244f-b0f2-452b-bcba-4c88e055c3ca